Add tests for common/util
