+++ packages/common/util.js
@@ -94,6 +94,13 @@ exports.omit = (objectIn, keys) => {
  * @param {Function} predicate - the predicate to negate
  * @returns {Function} the new negated function
  * @kind function
+ *
+ * @example
+ * const isEven = (x) => x % 2 === 0;
+ * const isOdd = negate(isEven);
+ *
+ * isOdd(2); // => false
+ * isOdd(3); // => true
  */
 exports.negate = (predicate) => (...args) => !predicate.apply(this, args);
